Refactored the baasaas client into separate files
