+++ integration-tests/baas-test-server/baasaas/assertions.ts
@@ -17,83 +17,21 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import assert from "node:assert";
-// import type {  } from "undici-types";
-
-const BAASAAS_BASE_URL = "https://us-east-1.aws.data.mongodb-api.com/app/baas-container-service-autzb/endpoint/";
-
-function createHeaders(authenticated = false) {
-  const headers: HeadersInit = {
-    "Content-Type": "application/json",
-  };
-  if (authenticated) {
-    const { BAASAAS_KEY } = process.env;
-    assert(BAASAAS_KEY, "Missing BAASAAS_KEY env");
-    headers["apiKey"] = BAASAAS_KEY;
-  }
-  return headers;
-}
-
-type RequestOptions = {
-  url: URL;
-  method: string;
-  authenticated: boolean;
-};
-
-async function request<R>(
-  { url, method, authenticated }: RequestOptions,
-  assertion: (response: unknown) => asserts response is R,
-): Promise<R> {
-  const response = await fetch(url, { method, headers: createHeaders(authenticated) });
-  if (response.ok) {
-    const json = await response.json();
-    assertion(json);
-    return json;
-  } else if (response.headers.get("content-type") === "application/json") {
-    const json = await response.json();
-    throw new Error(`Request failed (${response.status} / ${response.statusText}): ${JSON.stringify(json)}`);
-  } else {
-    throw new Error(`Request failed (${response.status} / ${response.statusText})`);
-  }
-}
 
 type StartedContainer = {
   id: string;
 };
 
-function assertStartedContainer(value: unknown): asserts value is StartedContainer {
+export function assertStartedContainer(value: unknown): asserts value is StartedContainer {
   assert(typeof value === "object" && value !== null);
-  assert("id" in value);
+  assert("id" in value, "Missing an id");
   assert.equal(typeof value.id, "string");
-}
-
-type StartContainerOptions =
-  | {
-      githash: string;
-    }
-  | {
-      branch: string;
-    };
-
-export async function startContainer(options: StartContainerOptions) {
-  const url = new URL("startContainer", BAASAAS_BASE_URL);
-  if ("githash" in options) {
-    url.searchParams.append("branch", options.githash);
-  } else if ("branch" in options) {
-    url.searchParams.append("branch", options.branch);
-  }
-  return request({ url, method: "POST", authenticated: true }, assertStartedContainer);
 }
 
 type StopContainerResponse = Record<never, never>;
 
-function assertStopContainerResponse(value: unknown): asserts value is StopContainerResponse {
+export function assertStopContainerResponse(value: unknown): asserts value is StopContainerResponse {
   /* nothing to assert - for now */
-}
-
-export async function stopContainer(id: string) {
-  const url = new URL("stopContainer", BAASAAS_BASE_URL);
-  url.searchParams.append("id", id);
-  return request({ url, method: "POST", authenticated: true }, assertStopContainerResponse);
 }
 
 type Container = {
@@ -108,7 +46,7 @@ type Container = {
   tags: unknown[];
 };
 
-function assertContainer(value: unknown): asserts value is Container {
+export function assertContainer(value: unknown): asserts value is Container {
   assert(typeof value === "object" && value !== null);
   assert("id" in value && typeof value.id === "string");
   assert("lastStatus" in value && typeof value.lastStatus === "string");
@@ -128,12 +66,6 @@ function assertContainer(value: unknown): asserts value is Container {
   assert("tags" in value && Array.isArray(value.tags));
 }
 
-export async function containerStatus(id: string) {
-  const url = new URL("containerStatus", BAASAAS_BASE_URL);
-  url.searchParams.append("id", id);
-  return request({ url, method: "GET", authenticated: true }, assertContainer);
-}
-
 type Userinfo = {
   id: string;
   type: string;
@@ -142,7 +74,7 @@ type Userinfo = {
   identities: [{ id: string; provider_type: string }];
 };
 
-function assertUserinfo(value: unknown): asserts value is Userinfo {
+export function assertUserinfo(value: unknown): asserts value is Userinfo {
   assert(typeof value === "object" && value !== null);
   assert("id" in value && typeof value.id === "string");
   assert("type" in value && typeof value.type === "string");
@@ -154,16 +86,6 @@ function assertUserinfo(value: unknown): asserts value is Userinfo {
     assert("id" in identity && typeof identity.id === "string");
     assert("provider_type" in identity && typeof identity.provider_type === "string");
   }
-}
-
-export async function userinfo() {
-  const url = new URL("userinfo", BAASAAS_BASE_URL);
-  const headers = createHeaders(true);
-  const response = await fetch(url, { headers });
-  assert(response.ok);
-  const json = await response.json();
-  assertUserinfo(json);
-  return json;
 }
 
 type Image = {
@@ -180,7 +102,7 @@ type Image = {
   imageTag: string;
 };
 
-function assertImage(value: unknown): asserts value is Image {
+export function assertImage(value: unknown): asserts value is Image {
   assert(typeof value === "object" && value !== null);
   assert("_id" in value && typeof value._id === "string");
   assert("project" in value && typeof value.project === "string");
@@ -200,9 +122,12 @@ type Images = {
   images: Record<string, undefined | Image[]>;
 };
 
-function assertImages(value: unknown): asserts value is Images {
+export function assertImages(value: unknown): asserts value is Images {
   assert(typeof value === "object" && value !== null);
   assert("allBranches" in value && Array.isArray(value.allBranches));
+  for (const branch of value.allBranches) {
+    assert.equal(typeof branch, "string");
+  }
   assert("images" in value && typeof value.images === "object" && value.images !== null);
   for (const available of Object.values(value.images)) {
     assert(Array.isArray(available));
@@ -212,24 +137,11 @@ function assertImages(value: unknown): asserts value is Images {
   }
 }
 
-export async function listImages() {
-  const url = new URL("images", BAASAAS_BASE_URL);
-  return request({ url, method: "GET", authenticated: false }, assertImages);
-}
-
 type ContainerList = Container[];
 
-function assertContainerList(value: unknown): asserts value is ContainerList {
+export function assertContainerList(value: unknown): asserts value is ContainerList {
   assert(Array.isArray(value));
   for (const element of value) {
     assertContainer(element);
   }
-}
-
-export async function listContainers(mine = false) {
-  const url = new URL("listContainers", BAASAAS_BASE_URL);
-  if (mine) {
-    url.searchParams.append("mine", "true");
-  }
-  return request({ url, method: "GET", authenticated: true }, assertContainerList);
 }
